pyproject.toml
[tool.poetry]
name = "py-aiger-coins"
version = "2.2.0"
description = "Library for creating circuits that encode discrete distributions."
authors = ["Marcell Vazquez-Chanlatte <mvc@linux.com>"]
repository = "https://github.com/mvcisback/py-aiger-coins"
readme="README.md"
license = "MIT"
packages = [
    { include = "aiger_coins"},
]

[tool.poetry.dependencies]
python = "^3.7"
attrs = "^20.0.0"
py-aiger = "^6.1.1"
py-aiger-bv = "^4.2.1"
funcy = "^1.13"
py-aiger-bdd = {version = "^3.0.0", optional = true}
py-aiger-sat = {version = "^3.0.4", optional = true}
py-aiger-discrete = "^0.1.2"

[tool.poetry.dev-dependencies]
codecov = "^2.0"
hypothesis = "^5.0.0"
py-aiger-bdd = "^3.0.0"
pytest-cov = "^2.7"
pytest-flake8 = "^1.0"
pytest-xdist = "^2.0.0"
py-aiger-ptltl = "^3.0.0"
python-sat = "^0.1.6-alpha.9"
py-aiger-sat = "^3.0.4"


[tool.poetry.extras]
bdd = ["py-aiger-bdd"]
sat = ["py-aiger-sat"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
